fix(blog): update blog to multi step form
