+++ src/Components/Form/MultiStep/MultiStepForm.css
@@ -1,0 +1,20 @@
+fieldset {   
+border: 1px groove #ddd !important;
+padding: 0 1.4em 1.4em 1.4em !important;
+margin: 0 0 1.5em 0 !important;
+-webkit-box-shadow:  0px 0px 0px 0px #000;
+box-shadow:  0px 0px 0px 0px #000;
+}
+
+legend {
+font-size: 1.2em !important;
+font-weight: bold !important;
+text-align: left !important;
+width:auto;
+padding:0 10px;
+border-bottom:none;
+}   
+
+.progress-bar {
+  transition: width 0.5s ease-in-out;
+}
